Made changes to have ReZound compile with gcc-3.1.1 and greatfully it still compiles the same under gcc-2.95.2 I properly fixed most of the problems I think. However, I had to work around one problem in src/PoolFile/TStaticPoolAccesser.h because I don't know the proper way to make a template parameter as a friend of the template And I did something that I don't think is necessary, but is just a bug and that is redefine 3 typedefs from TStaticPoolAccesser in the derived class TPoolAccesser I will have some further changes to make to the paste channels code to always use vector<vector<bool> > instead of bool** or bool[][]

+++ rezound/src/frontend_fox/CMainWindow.cpp
@@ -593,9 +593,9 @@ long CMainWindow::onShuttleChange(FXObject *sender,FXSelector sel,void *ptr)
 		if(shuttlePos==0)
 			playSpeed=1.0;
 		else if(shuttlePos>0)
-			playSpeed=(pow((float)shuttlePos/(float)maxValue,2.0)*100.0)+1.0;
+			playSpeed=(pow((double)shuttlePos/(double)maxValue,2.0)*100.0)+1.0;
 		else if(shuttlePos<0)
-			playSpeed=(pow((float)shuttlePos/(float)minValue,2.0)*-100.0)-1.0;
+			playSpeed=(pow((double)shuttlePos/(double)minValue,2.0)*-100.0)-1.0;
 
 		s->channel->setPlaySpeed(playSpeed);	
 	}
